+++ app/routes/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
 router.get('/', (req, res, next) => {
-    res.send('OK');
+  res.send('OK');
 });
 
 module.exports = router;
