+++ app/services/index.js
@@ -1,0 +1,5 @@
+const rainFallService = require('./rainFall.service');
+
+module.exports = {
+  rainFallService,
+};
